Fixes due to PR comments
